# ⚡️ perf: fix slow delete file sql